requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
transformers==4.38.0
torch==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
umap-learn==0.5.5
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.23.2 